Rename ThemeCustom to Theme
