+++ src/Core/Components/Theme/Styles/ThemeExtensions.cs
@@ -7,16 +7,16 @@ using System.Diagnostics.CodeAnalysis;
 namespace Microsoft.FluentUI.AspNetCore.Components;
 
 /// <summary />
-public static class ThemeCustomExtensions
+public static class ThemeExtensions
 {
     /// <summary>
     /// Converts the current theme custom object to a JSON string.
     /// </summary>
     /// <returns></returns>
     [RequiresUnreferencedCode("This method uses reflection which may be trimmed.")]
-    public static string ToJson(this ThemeCustom themeCustom)
+    public static string ToJson(this Theme theme)
     {
-        var dico = themeCustom.ToDictionary().Where(i => !string.IsNullOrEmpty(i.Value));
+        var dico = theme.ToDictionary().Where(i => !string.IsNullOrEmpty(i.Value));
         var data = System.Linq.Enumerable.ToDictionary(dico, x => x.Key, x => x.Value, StringComparer.Ordinal);
         var options = new System.Text.Json.JsonSerializerOptions
         {
@@ -28,39 +28,39 @@ public static class ThemeCustomExtensions
     /// <summary>
     /// Populates the theme custom object with values from a JSON string.
     /// </summary>
-    /// <param name="themeCustom"></param>
+    /// <param name="Theme"></param>
     /// <param name="json"></param>
     /// <returns></returns>
     [RequiresUnreferencedCode("This method uses reflection which may be trimmed.")]
-    public static ThemeCustom FromJson(this ThemeCustom themeCustom, string json)
+    public static Theme FromJson(this Theme Theme, string json)
     {
         var data = System.Text.Json.JsonSerializer.Deserialize<Dictionary<string, string?>>(json);
         if (data is not null)
         {
-            themeCustom.FromDictionary(data);
+            Theme.FromDictionary(data);
         }
 
-        return themeCustom;
+        return Theme;
     }
 }
 
-[SuppressMessage("Design", "MA0048:File name must match type name", Justification = "Extension methods for ThemeCustom should be in the same file as the class.")]
-public partial class ThemeCustom
+[SuppressMessage("Design", "MA0048:File name must match type name", Justification = "Extension methods for Theme should be in the same file as the class.")]
+public partial class Theme
 {
     /// <summary>
-    /// Initializes a new empty instance of the <see cref="ThemeCustom"/> class
+    /// Initializes a new empty instance of the <see cref="Theme"/> class
     /// </summary>
-    public ThemeCustom()
+    public Theme()
     {
 
     }
 
     /// <summary>
-    /// Initializes a new instance of the <see cref="ThemeCustom"/> class with values from a JSON string.
+    /// Initializes a new instance of the <see cref="Theme"/> class with values from a JSON string.
     /// </summary>
     /// <param name="json">The JSON string containing the theme custom values.</param>
     [RequiresUnreferencedCode("This constructor uses reflection which may be trimmed.")]
-    public ThemeCustom(string json)
+    public Theme(string json)
     {
         FromJson(json);
     }
@@ -70,7 +70,7 @@ public partial class ThemeCustom
     /// </summary>
     /// <returns></returns>
     [RequiresUnreferencedCode("This method uses reflection which may be trimmed.")]
-    public string ToJson() => ThemeCustomExtensions.ToJson(this);
+    public string ToJson() => ThemeExtensions.ToJson(this);
 
     /// <summary>
     /// Populates the theme custom object with values from a JSON string.
@@ -78,5 +78,5 @@ public partial class ThemeCustom
     /// <param name="json"></param>
     /// <returns></returns>
     [RequiresUnreferencedCode("This method uses reflection which may be trimmed.")]
-    public ThemeCustom FromJson(string json) => ThemeCustomExtensions.FromJson(this, json);
+    public Theme FromJson(string json) => ThemeExtensions.FromJson(this, json);
 }
